backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.45
psycopg[binary]>=3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
shapely==2.0.2
geopandas==0.14.2
osmnx==1.9.1
networkx==3.2.1
